backend/izzikg/src/requirements.txt
cent==3.0.1
Django==3.2
Pillow==8.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
# No imperative filtering, declarative beautiful filters
django-filter==2.4.0
# For process specification
django-fsm==2.7.1
djangorestframework==3.12.4
django-silk==4.1.0
django-3-jet==1.0.8
gunicorn==20.1.0
psycopg2-binary==2.8.6
pyjwt==2.0.1
requests==2.25.1
sentry-sdk==1.0.0
whitenoise==5.2.0
djangorestframework-simplejwt==4.6.0
django-ckeditor==6.0.0
django-solo==1.1.5
python-slugify==4.0.1
django-admin-sortable2==1.0
drf-yasg==1.20.0
qrcode==6.1
xmltodict==0.12.0
dicttoxml==1.7.4
django-2gis-maps==1.0.3
django-resized==0.3.11
drf-writable-nested==0.6.3
cryptography==39.0.0
